Make interrupt handlers configurable
Because the ISRs can only appear once in the crate graph, it is
important to allow downstream users to control their definition.  The
common way to make this configurable is via a default-on "rt" feature.

Define such a feature for this crate and guard the ISR definitions
behind it.  For downstream customizations, add two public functions
containing the ISR implementation that could be manually called.